# src/PaperMalKing.Startup/Web/InviteEligibility.cs
// SPDX-License-Identifier: AGPL-3.0-or-later
// Copyright (C) 2021-2026 N0D4N

namespace PaperMalKing.Startup.Web;

public enum InviteEligibility
{
	Allowed = 0,

	NotAllowed = 1,

	Unknown = 2,
}