# src/Tantei.Client/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "PaperMalKing | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/sign-in": {
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "SignIn",
        "parameters": [
          {
            "name": "returnUrl",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/sign-out": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "SignOut",
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "GetCurrentUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/config": {
      "get": {
        "tags": [
          "Config"
        ],
        "operationId": "GetSiteConfig",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteConfigResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/getUpdateTimes": {
      "get": {
        "tags": [
          "Status"
        ],
        "operationId": "GetUpdateTimes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UpdateProviderStatusResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/ping": {
      "get": {
        "tags": [
          "Status"
        ],
        "operationId": "Ping",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PingResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/guilds/manageable": {
      "get": {
        "tags": [
          "Guilds"
        ],
        "operationId": "GetManageableGuilds",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManageableGuildResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/guilds/invitable": {
      "get": {
        "tags": [
          "Guilds"
        ],
        "operationId": "GetInvitableGuilds",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitableGuildsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/guilds/{guildId}/invite": {
      "get": {
        "tags": [
          "Guilds"
        ],
        "operationId": "InviteToGuild",
        "parameters": [
          {
            "name": "guildId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "uint64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "302": {
            "description": "Found"
          },
          "403": {
            "description": "Forbidden"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/guilds/refresh": {
      "post": {
        "tags": [
          "Guilds"
        ],
        "operationId": "RefreshGuilds",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthStateResponse": {
        "required": [
          "user"
        ],
        "type": "object",
        "properties": {
          "user": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CurrentUserResponse"
              }
            ]
          }
        }
      },
      "CurrentUserResponse": {
        "required": [
          "discordUserId",
          "username",
          "avatarUrl",
          "isRegistered",
          "isWebAdmin"
        ],
        "type": "object",
        "properties": {
          "discordUserId": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "avatarUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "isRegistered": {
            "type": "boolean"
          },
          "isWebAdmin": {
            "type": "boolean"
          }
        }
      },
      "InvitableGuildResponse": {
        "required": [
          "guildId",
          "name",
          "iconUrl"
        ],
        "type": "object",
        "properties": {
          "guildId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "iconUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "InvitableGuildsResponse": {
        "required": [
          "eligibility",
          "guilds"
        ],
        "type": "object",
        "properties": {
          "eligibility": {
            "$ref": "#/components/schemas/InviteEligibility"
          },
          "guilds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvitableGuildResponse"
            }
          }
        }
      },
      "InviteEligibility": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "x-enum-varnames": [
          "Allowed",
          "NotAllowed",
          "Unknown"
        ]
      },
      "InviteMode": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "x-enum-varnames": [
          "Private",
          "SemiPrivate",
          "Public"
        ]
      },
      "ManageableGuildResponse": {
        "required": [
          "guildId",
          "name",
          "iconUrl"
        ],
        "type": "object",
        "properties": {
          "guildId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "iconUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "PingResponse": {
        "required": [
          "message",
          "timeUtc"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "timeUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "SiteConfigResponse": {
        "required": [
          "inviteMode"
        ],
        "type": "object",
        "properties": {
          "inviteMode": {
            "$ref": "#/components/schemas/InviteMode"
          }
        }
      },
      "UpdateProviderStatusResponse": {
        "required": [
          "name",
          "inProgress",
          "nextIn"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "inProgress": {
            "type": "boolean"
          },
          "nextIn": {
            "pattern": "^-?(\\d+\\.)?\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,7})?$",
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Auth"
    },
    {
      "name": "Config"
    },
    {
      "name": "Status"
    },
    {
      "name": "Guilds"
    }
  ]
}